Update create_input.Rmd
updated to create a list of rasters for my landscape.  cost function had to be scaled to 1/100 so I get comparable euclidean distance of ca. 1 as I had in my experiments. The model runs with the generated output, but somehow the species have a harder time surviving the initial phase.